backend/requirements.txt
annotated-types==0.6.0
bcrypt==3.2.0
certifi==2024.2.2
click==8.1.7
coloredlogs==15.0.1
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.3
flatbuffers==24.3.25
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanfriendly==10.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
loguru==0.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.16
mkdocs-material-extensions==1.3.1
mkdocs-swagger-ui-tag==0.7.1
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.17.1
opencv-python==4.9.0.80
orjson==3.10.3
packaging==24.0
pillow==10.3.0
priority==2.0.0
protobuf==5.26.1
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
scikit-learn==1.5.1
scipy==1.14.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
sympy==1.12
taskgroup==0.0.0a4
threadpoolctl==3.5.0
tomli==2.0.1
typer==0.12.3
typing_extensions==4.11.0
ujson==5.10.0
uvicorn==0.30.1
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
pyinstaller==6.11
pytest>=8.2
APScheduler==3.11.0
pre-commit>=2.21.0
black>=23.3.0
ruff>=0.0.241
psutil>=5.9.5
pytest-asyncio>=1.0.0
setuptools==66.1.1
transformers==4.39.0
faiss-cpu==1.8.0
ftfy==6.2.0
regex==2023.12.25
tokenizers==0.15.2
huggingface-hub==0.22.2
